requirements.txt
transformers>=4.34.0
torch>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
loguru>=0.7.2
pyyaml>=6.0.1